+++ src/prefect/tasks/kubernetes/job.py
@@ -170,6 +170,7 @@ class DeleteNamespacedJob(Task):
         namespace: str = "default",
         kube_kwargs: dict = None,
         kubernetes_api_key_secret: str = "KUBERNETES_API_KEY",
+        delete_option_kwargs: dict = None,
     ) -> None:
         """
         Task run method.
@@ -183,6 +184,9 @@ class DeleteNamespacedJob(Task):
             - kubernetes_api_key_secret (str, optional): the name of the Prefect Secret
                 which stored your Kubernetes API Key; this Secret must be a string and in
                 BearerToken format
+            - delete_option_kwargs (dict, optional): Optional keyword arguments to pass to
+                the V1DeleteOptions object (e.g. {"propagation_policy": "...",
+                "grace_period_seconds": "..."}.
 
         Raises:
             - ValueError: if `job_name` is `None`
@@ -207,11 +211,12 @@ class DeleteNamespacedJob(Task):
             api_client = client.BatchV1Api()
 
         kube_kwargs = {**self.kube_kwargs, **(kube_kwargs or {})}
+        delete_option_kwargs = delete_option_kwargs or {}
 
         api_client.delete_namespaced_job(
             name=job_name,
             namespace=namespace,
-            body=client.V1DeleteOptions(),
+            body=client.V1DeleteOptions(delete_option_kwargs),
             **kube_kwargs
         )